Restructure EventHandlerMethodInfo setup
- Moved the EventHandlerMethodInfo class to a separate file
- Unsure if the "CreateEventHandlerMethodInfo" method should be in the Dispatcher class. However, this is the only place where we are not using a generic TConfig, which makes usage a little easier. One could argue, however, that this method is better suited as a static method inside the EventHandlerMethodInfo class.
